Update Node and TernarySearchTree classes
Concerning the Node class, it's better to use this naming instead.  Also I've added a pointer to verify when a word has ended; 
For the TST class, only verifies if the word is empty and if a node needs to be created. 
I still need to build the verification part to attribute a node either to left, middle or right.