Added a place holder for the print function
Currently when I try to print the tree, it gives the place in memory rather than the tree itself, so I will make a function with the correct indentation and add it in the code. The printing function is not related to the insert but to the entire tree itself. 
To be done - pending